split up admin e2e
