+++ .github/workflows/main.yml
@@ -289,7 +289,8 @@ jobs:
         suite:
           - _community
           - access-control
-          - admin
+          - admin__e2e_1
+          - admin__e2e_2
           - auth
           - field-error-states
           - fields-relationship
